+++ packages/framer-motion/src/events/event-info.ts
@@ -1,64 +1,35 @@
 import { EventInfo } from "./types"
-import { isTouchEvent } from "../gestures/utils/event-type"
 
 /**
  * Filters out events not attached to the primary pointer (currently left mouse button)
  * @param eventHandler
  */
 function filterPrimaryPointer(eventHandler: EventListener): EventListener {
-    return (event: Event) => {
-        const isMouseEvent = event instanceof MouseEvent
-        const isPrimaryPointer =
-            !isMouseEvent ||
-            (isMouseEvent && (event as MouseEvent).button === 0)
-
-        if (isPrimaryPointer) {
+    return (event: PointerEvent) => {
+        if (event.isPrimary) {
             eventHandler(event)
         }
     }
 }
 
 export type EventListenerWithPointInfo = (
-    e: MouseEvent | TouchEvent | PointerEvent,
+    e: PointerEvent,
     info: EventInfo
 ) => void
 
-const defaultPagePoint = { pageX: 0, pageY: 0 }
-
-function pointFromTouch(e: TouchEvent, pointType: "page" | "client" = "page") {
-    const primaryTouch = e.touches[0] || e.changedTouches[0]
-    const point = primaryTouch || defaultPagePoint
-
-    return {
-        x: point[pointType + "X"],
-        y: point[pointType + "Y"],
-    }
-}
-
-function pointFromMouse(
-    point: MouseEvent | PointerEvent,
-    pointType: "page" | "client" = "page"
-) {
-    return {
-        x: point[pointType + "X"],
-        y: point[pointType + "Y"],
-    }
-}
-
 export function extractEventInfo(
-    event: MouseEvent | TouchEvent | PointerEvent,
+    event: PointerEvent,
     pointType: "page" | "client" = "page"
 ): EventInfo {
     return {
-        point: isTouchEvent(event)
-            ? pointFromTouch(event, pointType)
-            : pointFromMouse(event, pointType),
+        point: {
+            x: event[pointType + "X"],
+            y: event[pointType + "Y"],
+        },
     }
 }
 
-export function getViewportPointFromEvent(
-    event: MouseEvent | TouchEvent | PointerEvent
-) {
+export function getViewportPointFromEvent(event: PointerEvent) {
     return extractEventInfo(event, "client")
 }