+++ packages/framer-motion/src/events/use-pointer-event.ts
@@ -1,52 +1,6 @@
 import { RefObject } from "react"
 import { useDomEvent, addDomEvent } from "./use-dom-event"
 import { wrapHandler, EventListenerWithPointInfo } from "./event-info"
-import {
-    supportsPointerEvents,
-    supportsTouchEvents,
-    supportsMouseEvents,
-} from "./utils"
-
-interface PointerNameMap {
-    pointerdown: string
-    pointermove: string
-    pointerup: string
-    pointercancel: string
-    pointerover?: string
-    pointerout?: string
-    pointerenter?: string
-    pointerleave?: string
-}
-
-const mouseEventNames: PointerNameMap = {
-    pointerdown: "mousedown",
-    pointermove: "mousemove",
-    pointerup: "mouseup",
-    pointercancel: "mousecancel",
-    pointerover: "mouseover",
-    pointerout: "mouseout",
-    pointerenter: "mouseenter",
-    pointerleave: "mouseleave",
-}
-
-const touchEventNames: PointerNameMap = {
-    pointerdown: "touchstart",
-    pointermove: "touchmove",
-    pointerup: "touchend",
-    pointercancel: "touchcancel",
-}
-
-function getPointerEventName(name: string): string {
-    if (supportsPointerEvents()) {
-        return name
-    } else if (supportsTouchEvents()) {
-        return touchEventNames[name]
-    } else if (supportsMouseEvents()) {
-        return mouseEventNames[name]
-    }
-
-    return name
-}
 
 export function addPointerEvent(
     target: EventTarget,
@@ -56,7 +10,7 @@ export function addPointerEvent(
 ) {
     return addDomEvent(
         target,
-        getPointerEventName(eventName),
+        eventName,
         wrapHandler(handler, eventName === "pointerdown"),
         options
     )
@@ -70,7 +24,7 @@ export function usePointerEvent(
 ) {
     return useDomEvent(
         ref,
-        getPointerEventName(eventName),
+        eventName,
         handler && wrapHandler(handler, eventName === "pointerdown"),
         options
     )