+++ packages/framer-motion/src/gestures/__tests__/hover.test.tsx
@@ -1,7 +1,11 @@
-import { mouseEnter, mouseLeave, render } from "../../../jest.setup"
+import {
+    pointerDown,
+    pointerEnter,
+    pointerLeave,
+    render,
+} from "../../../jest.setup"
 import * as React from "react"
 import { motion } from "../../"
-import { fireEvent } from "@testing-library/react"
 import { motionValue } from "../../value"
 import { transformValues } from "../../motion/__tests__/util-transform-values"
 import { sync } from "../../frameloop"
@@ -15,8 +19,8 @@ describe("hover", () => {
         )
 
         const { container } = render(<Component />)
-        mouseEnter(container.firstChild as Element)
-        mouseLeave(container.firstChild as Element)
+        pointerEnter(container.firstChild as Element)
+        pointerLeave(container.firstChild as Element)
 
         expect(hoverIn).toBeCalledTimes(1)
         expect(hoverOut).toBeCalledTimes(1)
@@ -36,7 +40,7 @@ describe("hover", () => {
             const { container, rerender } = render(<Component />)
             rerender(<Component />)
 
-            mouseEnter(container.firstChild as Element)
+            pointerEnter(container.firstChild as Element)
 
             resolve(opacity.get())
         })
@@ -63,7 +67,7 @@ describe("hover", () => {
             const { container, rerender } = render(<Component />)
             rerender(<Component />)
 
-            mouseEnter(container.firstChild as Element)
+            pointerEnter(container.firstChild as Element)
 
             resolve(opacity.get())
         })
@@ -99,7 +103,7 @@ describe("hover", () => {
 
             const { container } = render(<Component />)
 
-            mouseEnter(container.firstChild as Element)
+            pointerEnter(container.firstChild as Element)
             resolve(opacity.get())
         })
 
@@ -130,10 +134,10 @@ describe("hover", () => {
                 <Component onAnimationComplete={onComplete} />
             )
 
-            mouseEnter(container.firstChild as Element)
+            pointerEnter(container.firstChild as Element)
             setTimeout(() => {
                 hasMousedOut = true
-                mouseLeave(container.firstChild as Element)
+                pointerLeave(container.firstChild as Element)
             }, 10)
         })
 
@@ -161,8 +165,8 @@ describe("hover", () => {
             const { container, rerender } = render(<Component />)
             rerender(<Component />)
 
-            fireEvent.mouseDown(container.firstChild as Element)
-            mouseEnter(container.firstChild as Element)
+            pointerDown(container.firstChild as Element)
+            pointerEnter(container.firstChild as Element)
 
             resolve([opacity.get(), scale.get()])
         })
@@ -188,10 +192,10 @@ describe("hover", () => {
             const { container, rerender } = render(<Component />)
             rerender(<Component />)
 
-            mouseEnter(container.firstChild as Element)
+            pointerEnter(container.firstChild as Element)
 
             sync.postRender(() => {
-                mouseLeave(container.firstChild as Element)
+                pointerLeave(container.firstChild as Element)
                 sync.postRender(() => resolve(container.firstChild as Element))
             })
         })