+++ packages/framer-motion/src/gestures/use-hover-gesture.ts
@@ -1,4 +1,3 @@
-import { isMouseEvent } from "./utils/event-type"
 import { EventInfo } from "../events/types"
 import { AnimationType } from "../render/utils/types"
 import { usePointerEvent } from "../events/use-pointer-event"
@@ -6,12 +5,16 @@ import type { VisualElement } from "../render/VisualElement"
 import { FeatureProps } from "../motion/features/types"
 import { isDragActive } from "./drag/utils/lock"
 
+function isMouseEvent(event: PointerEvent) {
+    return event.type !== "pen" && event.type !== "touch"
+}
+
 function createHoverEvent(
     visualElement: VisualElement,
     isActive: boolean,
-    callback?: (event: MouseEvent, info: EventInfo) => void
+    callback?: (event: PointerEvent, info: EventInfo) => void
 ) {
-    return (event: MouseEvent, info: EventInfo) => {
+    return (event: PointerEvent, info: EventInfo) => {
         if (!isMouseEvent(event) || isDragActive()) return
 
         /**